Use delegation rather than inheritance to decouple JLine
Fixes #78
